+++ packages/Ecotone/src/Lite/Test/Configuration/DelayedMessageReleaseHandler.php
@@ -9,7 +9,6 @@ use Ecotone\Messaging\Channel\DelayableQueueChannel;
 use Ecotone\Messaging\Channel\MessageChannelInterceptorAdapter;
 use Ecotone\Messaging\Handler\ChannelResolver;
 use Ecotone\Messaging\Scheduling\TimeSpan;
-use Ecotone\Messaging\Support\Assert;
 
 /**
  * licence Apache-2.0
@@ -18,7 +17,7 @@ final class DelayedMessageReleaseHandler
 {
     public function releaseMessagesAwaitingFor(string $channelName, int|TimeSpan|DateTimeInterface $timeInMillisecondsOrDateTime, ChannelResolver $channelResolver): void
     {
-        if (!$channelResolver->hasChannelWithName($channelName)) {
+        if (! $channelResolver->hasChannelWithName($channelName)) {
             return;
         }